All errors resolved, Build Successful

+++ src/entry.c
@@ -874,7 +874,7 @@ struct stumpless_entry *
 stumpless_set_entry_message_str_w( struct stumpless_entry *entry,
                                    const wchar_t *message ) {
   char *new_message;
-  size_t new_message_length;
+  int new_message_length;
   const char *old_message;
 
   VALIDATE_ARG_NOT_NULL( entry );
